Experimental eslint flat config, update ext in configs
