+++ package.json
@@ -22,8 +22,8 @@
     "prepare": "husky install",
     "prettier": "prettier --check \"**/*.+(js|cjs|jsx|ts|tsx|pcss)\"",
     "prettier:fix": "prettier --write \"**/*.+(js|cjs|jsx|ts|tsx|pcss)\"",
-    "lint": "eslint src --ext .js,.cjs,.jsx,.ts,.tsx",
-    "lint:fix": "eslint src --ext .js,.cjs,.jsx,.ts,.tsx --fix",
+    "lint": "eslint **/*{js,mjs,cjs,jsx,mjsx,ts,mts,cts,tsx,mtsx}",
+    "lint:fix": "eslint **/*{js,mjs,cjs,jsx,mjsx,ts,mts,cts,tsx,mtsx} --fix",
     "lint:fix-staged": "eslint --fix",
     "lint:css": "stylelint src",
     "lint:css:fix": "pnpm lint:css -- --fix",
@@ -53,8 +53,6 @@
     "dotenv-webpack": "^8.0.1",
     "esbuild-loader": "^2.20.0",
     "eslint": "^8.30.0",
-    "eslint-config-airbnb": "^19.0.4",
-    "eslint-config-airbnb-typescript": "^17.0.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-import-resolver-typescript": "^3.5.2",
     "eslint-import-resolver-webpack": "^0.13.2",
@@ -64,6 +62,7 @@
     "eslint-plugin-react": "^7.31.11",
     "eslint-plugin-react-hooks": "^4.6.0",
     "fork-ts-checker-webpack-plugin": "^7.2.14",
+    "globals": "^13.19.0",
     "html-webpack-plugin": "^5.5.0",
     "husky": "^8.0.2",
     "lint-staged": "^13.1.0",
