+++ tsconfig.json
@@ -1,11 +1,15 @@
 {
   "include": [
-    ".eslintrc.cjs",
-    "./**/*.ts",
-    "./**/*.tsx",
     "./**/*.js",
     "./**/*.cjs",
-    "./**/*.jsx"
+    "./**/*.mjs",
+    "./**/*.jsx",
+    "./**/*.mjsx",
+    "./**/*.ts",
+    "./**/*.cts",
+    "./**/*.mts",
+    "./**/*.tsx",
+    "./**/*.mtsx",
   ],
   "exclude": [
     "node_modules"
